refactor: Iniciada alteração arquitetural do sistema para criar o serviço de compartilhamento de biometrias independente, a nível de implementação - Isso permitiria à empresa uma venda independente deste serviço, além de ritmos de desenvolvimento e entrega diferentes das demais funcionalidades do sistema; - Para que seja referenciado nas 2 aplicações, criado projeto de 'Domain'; - Por entender que esses objetos são como entidades para a ação a ser realizada pelo código de interação com API externa, repassados DTOs antes em diretório de 'Application', para projeto de 'Domain';     - Atualmente, é o objeto exposto na camada de 'Application', porém, se um usuário solicitasse alguma informação que não é de responsabilidade do equipamento (mas do sistema), seria criado um objeto com este propósito em 'Application';
