+++ CleanArchitecture.ISP.Domain/Brands/Entities/Brand.cs
