+++ CleanArchitecture.ISP.Domain/CleanArchitecture.ISP.Domain.csproj
@@ -1,14 +1,9 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\CleanArchitecture.ISP.Domain\CleanArchitecture.ISP.Domain.csproj" />
-  </ItemGroup>
 
 </Project>
