+++ CleanArchitecture.ISP.Domain/OperationResult/SetDateTime.cs
@@ -1,0 +1,5 @@
+﻿namespace CleanArchitecture.ISP.Domain.OperationResult;
+public record SetDateTimeSuccess(string PointClockIp) : Result(true)
+{
+    public DateTime SettedDateTime => DateTime.UtcNow;
+}
