+++ CleanArchitecture.ISP.Domain/PointClocks/Entities/PointClock.cs
