+++ CleanArchitecture.ISP/Application/UseCases/ShareBiometrics.cs
@@ -1,6 +1,7 @@
-﻿using CleanArchitecture.ISP.Application.OperationResult;
+﻿using CleanArchitecture.ISP.Domain.OperationResult;
 using CleanArchitecture.ISP.Domain.PointClocks.Entities;
 using CleanArchitecture.ISP.Domain.PointClocks.Gateway;
+using CleanArchitecture.ISP.Infrastructure.PointClockGateways;
 
 namespace CleanArchitecture.ISP.Application.UseCases;
 public class ShareBiometrics
@@ -29,19 +30,4 @@ public class ShareBiometrics
 
 public record ShareBiometricsRequest(PointClock[] PointClocks)
 {
-}
-
-public record ShareBiometricsSuccess(string PointClockIp, int NumberOfSharedBiometrics) : Result(true)
-{
-}
-
-public record ShareBiometricsPointClockNotSupported : Result
-{
-    public string MessageError { get; }
-
-    public ShareBiometricsPointClockNotSupported(string messageError)
-        : base(false)
-    {
-        MessageError = "The functionality was not executed. Reason: " + messageError;
-    }
 }
