+++ CleanArchitecture.ISP/Program.cs
@@ -1,7 +1,7 @@
 ﻿using CleanArchitecture.ISP;
-using CleanArchitecture.ISP.Application.OperationResult;
 using CleanArchitecture.ISP.Application.UseCases;
 using CleanArchitecture.ISP.Domain.Brands.Entities;
+using CleanArchitecture.ISP.Domain.OperationResult;
 using CleanArchitecture.ISP.Domain.PointClocks.Entities;
 
 PointClock[] pointClocks = [
